增加开关

+++ init.c
@@ -7,43 +7,56 @@
 #include "dpd_err.h"
 
 /* dpd信号复位 */
-void dpd_reset(void)
+static void dpd_reset(int i)
 {
     system("devmem 0x40000480 32 0x0");
+    if (i == 0)
+    {
+        system("killall /usr/bin/dpd-smp");
+    }
     sleep(0.5);
     system("devmem 0x40000480 32 0x1");
 }
 
 /* dpd初始化 */
-void dpd_init(void)
+void dpd_init(int tmp)
 {
-    int i, k;
-    for (k = 0; k < 3; k++)
+    if (tmp == 1)
     {
-        dpd_reset();
-        i = CODEPOINTER(125);
-        if (i == 2)
+        int i, k;
+        for (k = 0; k < 3; k++)
         {
-            system("/usr/bin/dpd-smp -u 0 &");
-            i = CODEPOINTER(130);
-            switch (i)
+            dpd_reset(1);
+            i = CODEPOINTER(125);
+            if (i == 2)
             {
-            case 3:
-                i = CODEPOINTER(128);
+                system("/usr/bin/dpd-smp -u 0 &");
+                i = CODEPOINTER(130);
                 switch (i)
                 {
-                case 4:
-                    fprintf(stdout, "DPD init successful.\n");
-                    k = 3;
+                case 3:
+                    i = CODEPOINTER(128);
+                    switch (i)
+                    {
+                    case 4:
+                        fprintf(stdout, "DPD init successful.\n");
+                        k = 3;
+                        break;
+                    default:
+                        break;
+                    }
                     break;
                 default:
+                    system("killall /usr/bin/dpd-smp");
                     break;
                 }
-                break;
-            default:
-                system("killall /usr/bin/dpd-smp");
-                break;
             }
         }
+    }
+    if (tmp == 0)
+    {
+        dpd_reset(0);
+        dpd_init(1);
+        fprintf(stdout, "reset successful.");
     }
 }
